pyproject.toml
[project]
name = "etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "pre-commit>=4.2.0",
    "psycopg[binary]>=3.2.9",
    "pylint>=3.3.7",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "requests-mock>=1.12.1",
]
